+++ docs/src/release_notes.rst
@@ -5,6 +5,8 @@ Release notes
 =============
 3.5.1 (2021-02-12)
 ------------------
+Fixed
+*****
 - :ref:`client` - document decreased limit of :meth:`search <Spotify.search>`
   total (:issue:`241`)